requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.11.8
chardet==3.0.4
decorator==4.4.2
Django==1.8
django-celery==3.3.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
kombu==3.0.37
moviepy==1.0.3
numpy==1.19.3
Pillow==8.0.1
proglog==0.1.9
pytz==2020.4
requests==2.25.0
tqdm==4.51.0
urllib3==1.26.2
